Fix numpy output serialization in KeypointOutputProcessor and embed_all

+++ fiftyone/utils/torch.py
@@ -1617,7 +1617,7 @@ class KeypointOutputProcessor(OutputProcessor):
                     # Low confidence
                     points.append((float("nan"), float("nan")))
                 else:
-                    points.append((p[0] / width, p[1] / height))
+                    points.append((float(p[0] / width), float(p[1] / height)))
             _keypoints.append(
                 fol.Keypoint(
                     points=points,
